feat(analysis): started rewriting analysis with a dynamic layout
