+++ package.json
@@ -12,6 +12,8 @@
     "format": "prettier --write './src/**/*.{js,jsx,ts,tsx,css,md,json}' --config ./.prettierrc"
   },
   "dependencies": {
+    "@dnd-kit/core": "^6.3.1",
+    "@dnd-kit/sortable": "^10.0.0",
     "@hookform/resolvers": "^3.9.0",
     "@radix-ui/react-avatar": "^1.1.1",
     "@radix-ui/react-dialog": "^1.1.2",
@@ -42,11 +44,14 @@
     "next-themes": "^0.3.0",
     "react": "^18.3.1",
     "react-day-picker": "^8.10.1",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.3.1",
     "react-flagkit": "^2.0.4",
     "react-hook-form": "^7.53.1",
     "react-hotkeys-hook": "^4.6.1",
     "react-i18next": "^15.1.1",
+    "react-resizable-panels": "^2.1.7",
     "react-router-dom": "^6.26.2",
     "recharts": "^2.13.0",
     "sonner": "^1.5.0",
@@ -59,6 +64,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.12.0",
+    "@iconify/react": "^5.1.0",
     "@types/node": "^22.7.9",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
