+++ src/pages/analysis/panels/engineInterpretation/engine-interpretation.tsx
@@ -1,0 +1,5 @@
+﻿export const EngineInterpretation = () => (
+  <div className="flex h-full items-center justify-center p-6">
+    <span className="font-semibold">Engine Interpretation</span>
+  </div>
+);
