+++ src/pages/analysis/panels/engineLines/engine-lines.tsx
@@ -1,0 +1,5 @@
+﻿export const EngineLines = () => (
+  <div className="flex h-full items-center justify-center p-6">
+    <span className="font-semibold">Engine Lines</span>
+  </div>
+);
