+++ src/pages/analysis/panels/moveList/move-list.tsx
@@ -1,0 +1,5 @@
+﻿export const MoveList = () => (
+  <div className="flex h-full items-center justify-center p-6">
+    <span className="font-semibold">Move List</span>
+  </div>
+);
